upadte ngày 27/2/2023

+++ BE/src/main/java/com/longkubi/qlns/model/dto/TimeKeepingDto.java
@@ -2,16 +2,11 @@ package com.longkubi.qlns.model.dto;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
-import com.longkubi.qlns.model.entity.CommendationAndDiscipline;
-import com.longkubi.qlns.model.entity.Employee;
-import com.longkubi.qlns.model.entity.PaymentSalary;
-import io.swagger.models.auth.In;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.Column;
 import java.util.Date;
 import java.util.UUID;
 
@@ -23,40 +18,22 @@ import java.util.UUID;
 public class TimeKeepingDto {
 
     private UUID id;
-
     private String code;//mã chấm công
-
     private EmployeeDto employee;//nhân viên
-
-    private Integer numberWorkDay;//Số Ngày Công Đi Làm
-
-    private Integer status;//
-
+    private Byte numberWorkDay;//Số Ngày Công Đi Làm
+    private Byte status;//
 
     //    @Column(name = "_leave")
     //    private int leave;
-
-    private Integer numberDayOff;//Số Ngày Nghỉ Có Phép
-
-    private Integer numberDayUnexcusedLeave;//Số Ngày Nghỉ Không Phép
-
-    private Integer month;//tháng
-
-    private Integer year;//năm
-
-    private Integer numberOvertimeHours; //Giờ Làm Thêm
-
+    private Byte numberDayOff;//Số Ngày Nghỉ Có Phép
+    private Byte numberDayUnexcusedLeave;//Số Ngày Nghỉ Không Phép
+    private Byte month;//tháng
+    private Short year;//năm
+    private Short numberOvertimeHours; //Giờ Làm Thêm
     private String creator;
-
     private Date dateCreated;
-
-
     private String changedBy;
-
-
     private Date dateChange;
-
     //private CommendationAndDisciplineDto commendationAndDiscipline;
-
     private PaymentSalaryDto paymentSalary;
 }
